[refactor] domain package 변경
